# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "GET"
                ],
                "summary": "Get all post",
                "description": "Get title and description",
                "operationId": "8c292c8777126dc6d3efb05793ca8517",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "POST"
                ],
                "summary": "Create post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Post Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is a new post description"
                                    }
                                },
                                "required": [
                                    "title",
                                    "description"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },

        "/api/posts/{id}": {
            "put": {
                "tags": [
                    "Put"
                ],
                "summary": "Update a post by ID",
                "description": "Update a post with the provided data by its unique identifier",
                "operationId": "updatePostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the post to update",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Post Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is an updated post description"
                                    }
                                },
                                "required": [
                                    "title",
                                    "description"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },

                "delete": {
                    "tags": [
                        "Delete"
                    ],
                    "summary": "Delete a post by ID",
                    "description": "Delete a post with the provided ID",
                    "operationId": "deletePostById",
                    "parameters": [
                        {
                            "name": "id",
                            "in": "path",
                            "required": true,
                            "description": "ID of the post to delete",
                            "schema": {
                                "type": "integer"
                            }
                        }
                    ],
                    "responses": {
                        "200": {
                            "description": "OK",
                            "content": {
                                "application/json": {}
                            }
                        },
                        "404": {
                            "description": "Post not found",
                            "content": {
                                "application/json": {}
                            }
                        }
                    }
                }
            },

            "/api/users": {
                "get": {
                    "tags": [
                        "GETUsers"
                    ],
                    "summary": "Get all users",
                    "operationId": "c457726701591d1183b53aa71fc13441",
                    "responses": {
                        "200": {
                            "description": "All Users"
                        },
                        "400": {
                            "description": "Bad request"
                        },
                        "404": {
                            "description": "Resource Not Found"
                        }
                    },
                    "security": [
                        {
                            "bearerAuth": []
                        }
                    ]
                },
                "post": {
                    "tags": [
                        "PostUsers"
                    ],
                    "summary": "Create a new user",
                    "description": "Create a new user with the provided username, email, and password",
                    "operationId": "592819a0265360b2014512d6dbfaf0e7",
                    "requestBody": {
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "name",
                                        "email",
                                        "password"
                                    ],
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "john_doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@example.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Create New User"
                        },
                        "400": {
                            "description": "Bad request"
                        },
                        "404": {
                            "description": "Resource Not Found"
                        }
                    }
                }
            },
            "/api/users/{id}": {
                "get": {
                    "tags": [
                        "GetDetailUsers"
                    ],
                    "summary": "Get a specific user",
                    "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                    "parameters": [
                        {
                            "name": "id",
                            "in": "path",
                            "description": "ID of the user to retrieve",
                            "required": true,
                            "schema": {
                                "type": "integer"
                            }
                        }
                    ],
                    "responses": {
                        "200": {
                            "description": "User Detail"
                        },
                        "400": {
                            "description": "Bad request"
                        },
                        "404": {
                            "description": "Resource Not Found"
                        }
                    },
                    "security": [
                        {
                            "bearerAuth": []
                        }
                    ]
                },
                "delete": {
                    "tags": [
                        "DeleteUsers"
                    ],
                    "summary": "Delete a specific user",
                    "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                    "parameters": [
                        {
                            "name": "id",
                            "in": "path",
                            "description": "User ID",
                            "required": true,
                            "schema": {
                                "type": "integer"
                            }
                        }
                    ],
                    "responses": {
                        "200": {
                            "description": "Delate user"
                        }
                    },
                    "security": [
                        {
                            "bearerAuth": []
                        }
                    ]
                },
                "put": {
                    "tags": [
                        "PutUsers"
                    ],
                    "summary": "Update a specific user",
                    "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                    "parameters": [
                        {
                            "name": "id",
                            "in": "path",
                            "description": "User ID",
                            "required": true,
                            "schema": {
                                "type": "integer"
                            }
                        }
                    ],
                    "requestBody": {
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "john_doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@example.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Update User"
                        }
                    },
                    "security": [
                        {
                            "bearerAuth": []
                        }
                    ]
                }
            }

}
}